MenuScene fixed, added map selection
Menu scene navigation fixed, now you can properly navigate the menu without breaking things. This means that adding teams and maps to the selection screen should be easier than before, they are just a button with an OnClick method attached. This method is always pretty much the same, you should just add an element to the respective array and reference it when adding the method