Turn system added, "destroy ball" interaction modified
Turn system that accounts for a max number of action per turn, state machine implemented. I had to modify the "destroy" interaction that we used: since FindGameObjectsWithTag returns an array we couldn't destroy the gameobjects associated with its elements. So instead of destroying the gameobject we can now just disable its components using the DisableBall method.