Back to menu from level functionality, fix to lewd team and tutorial
You can go back to the menu by pressing the spacebar and answer yes to the prompt. The info message that informs the players on what button to press disappears after 3.5 seconds from the start of the scene. The time can be reduced in code
Lewd team stat fixed too